Module Cut_form.Pos

val at : t -> Logtk.Position.t -> term

Return the subterm at the given position, or

raises Invalid_argument

if the position is not valid

val lit_at : t -> Logtk.Position.t -> Logtk.Literal.t * Logtk.Position.t

Lookup which literal the position is about, return it and the rest of the position.

raises Invalid_argument

if the position is not valid

val clause_at : t -> Logtk.Position.t -> clause * Logtk.Position.t

Lookup which clause the position is about, return it and the rest of the position.

raises Invalid_argument

if the position is not valid

val replace : t -> at:Logtk.Position.t -> by:term -> t

In-place modification of the array, in which the subterm at given position is replaced by the by term.

raises Invalid_argument

if the position is not valid

val replace_many : t -> term Logtk.Position.Map.t -> t

In-place modification of the array, in which the subterm at given position is replaced by the by term.

raises Invalid_argument

if the position is not valid