Module S.WeightFun

Weight functions

type t = Stm.t -> int

attribute a weight to a stream. The smaller, the better (lightweight streams will be favored). A weight must always be positive.

val penalty : t

Returns the penalty of the stream

val combine : (t * int) list -> t

Combine a list of pairs w, coeff where w is a weight function, and coeff a strictly positive number. This is a weighted sum of weights.