Module ProofState.PassiveSet

include CLAUSE_SET

val on_add_clause : C.t Logtk.Signal.t

signal triggered when a clause is added to the set

val on_remove_clause : C.t Logtk.Signal.t

signal triggered when a clause is removed from the set

val add : C.t Iter.t -> unit

Add clauses to the set

val remove : C.t Iter.t -> unit

Remove clauses from the set

val clauses : unit -> C.ClauseSet.t

Current set of clauses

val is_passive : C.t -> bool

is the given clause passive

val queue : CQueue.t

Current state of the clause queue

val next : unit -> C.t option

Get-and-remove the next passive clause to process

val num_clauses : unit -> int