Module type PS.CLAUSE_SET

val on_add_clause : C.t Logtk.Signal.t

signal triggered when a clause is added to the set

val on_remove_clause : C.t Logtk.Signal.t

signal triggered when a clause is removed from the set

val add : C.t Iter.t -> unit

Add clauses to the set

val remove : C.t Iter.t -> unit

Remove clauses from the set