Module Libzipperposition_calculi.Simplex

module type OrderedType = sig ... end

The types of the variables used by the equations to solve

module type S = sig ... end

module Make : functor (Var : OrderedType) -> S with type var = Var.t

Functor building an implementation of the simplex solver given a totally ordered type for the variables

Higher-Level Interface

module type HELPER = sig ... end

module MakeHelp : functor (Var : OrderedType) -> HELPER with type external_var = Var.t