Module SubsumptionIndex.C

type t = C.t

val compare : t -> t -> int

Compare two clauses

val to_lits : t -> Logtk.Index_intf.lits

Iterate on literals of the clause

val labels : t -> Logtk.Index_intf.labels

Some integer labels. We assume that if c to subsume d, then labels c is a subset of labels d