Module ProofState.TermIndex

type t

type elt = C.WithPos.t

module Leaf : Logtk.Index_intf.LEAF with type elt = elt

val name : string

val empty : unit -> t

val is_empty : t -> bool

val size : t -> int

val add : t -> Logtk.Index_intf.term -> elt -> t

val add_seq : t -> (Logtk.Index_intf.term * elt) Iter.t -> t

val add_list : t -> (Logtk.Index_intf.term * elt) list -> t

val remove : t -> Logtk.Index_intf.term -> elt -> t

val update_leaf : t -> Logtk.Index_intf.term -> (elt -> bool) -> t

val remove_seq : t -> (Logtk.Index_intf.term * elt) Iter.t -> t

val remove_list : t -> (Logtk.Index_intf.term * elt) list -> t

val iter : t -> (Logtk.Index_intf.term -> elt -> unit) -> unit

val fold : t -> ('a -> Logtk.Index_intf.term -> elt -> 'a) -> 'a -> 'a

val retrieve_unifiables : t Logtk.Scoped.t -> Logtk.Index_intf.term Logtk.Scoped.t -> (Logtk.Index_intf.term * elt * Logtk.Unif_subst.t) Iter.t

Retrieves a decidable fragment of unifiables. Only one unifier per subterm.

val retrieve_unifiables_complete : ?⁠unif_alg:(Logtk.Term.t Logtk.Scoped.t -> Logtk.Term.t Logtk.Scoped.t -> Logtk.Unif_subst.t option OSeq.t) -> t Logtk.Scoped.t -> Logtk.Index_intf.term Logtk.Scoped.t -> (Logtk.Index_intf.term * elt * Logtk.Unif_subst.t option OSeq.t) Iter.t

Retrieves all unifiables. The set of unifiers is potentially infinite. Because HO unification is undecidable, the sequence is intersperced with `None`s to ensure termination for each element of the sequence.

val retrieve_generalizations : ?⁠subst:Logtk.Index_intf.subst -> t Logtk.Scoped.t -> Logtk.Index_intf.term Logtk.Scoped.t -> (Logtk.Index_intf.term * elt * Logtk.Index_intf.subst) Iter.t

val retrieve_specializations : ?⁠subst:Logtk.Index_intf.subst -> t Logtk.Scoped.t -> Logtk.Index_intf.term Logtk.Scoped.t -> (Logtk.Index_intf.term * elt * Logtk.Index_intf.subst) Iter.t

val to_dot : elt CCFormat.printer -> t CCFormat.printer

print oneself in DOT into the given file